Update fan service to support auto and manual modes.
Modify the code so that we can support automatic and manual modes for
the fan controller.  Automatic should be used to control the
temperatures unless there are extenuating circumstances, like having
SSD's of other heat generators in the system along with the raspberry
pi.  Using manual mode every CPU temperature can have a specific fan
speed.